requirements.txt
flask
flask-socketio
numpy
peewee
pillow
